+++ integration_tests/models/schema_tests/test_recency.sql
@@ -1,12 +1,2 @@
-
-{% if target.type == 'postgres' %}
-
 select
     {{ dbt_utils.date_trunc('day', dbt_utils.current_timestamp()) }} as today
-
-{% else %}
-
-select
-    cast({{ dbt_utils.date_trunc('day', dbt_utils.current_timestamp()) }} as datetime) as today
-    
-{% endif %}