+++ macros/schema_tests/recency.sql
@@ -12,6 +12,6 @@ select
     end as error_result
 from {{model}}
 where {{column_name}} >=
-    {{dbt_utils.dateadd(datepart, interval * -1, dbt_utils.current_timestamp())}}
+    {{ dbt_utils.dateadd(datepart, interval * -1, dbt_utils.current_timestamp()) }}
 
 {% endmacro %}